requirements.txt
numpy
opencv-python
torch==1.1.0
torchvision==0.3.0
tqdmpip
flask=2.1.0